README.md
Based on UMichigan's coursera Intro to DS with Python course.

Repository includes solutions for Week 2 to Week 4 assignments of Course 1 and Solutions to all assignments of Course 2; 

For Course 1, Week 1 was a quiz and hence solutions are not provided. 

I've found reading through other people's work a lot of help and unfortunately solutions available online are not coherent with what is required in assignments. Posting answers so that you can save hours of irritation and submitting wrong solutions.
Hope this helps!